+++ charts/internal/crds-firewall/templates/firewall-controller-manager/firewall.metal-stack.io_firewalls.yaml
@@ -67,19 +67,20 @@ spec:
             description: Spec contains the firewall specification.
             properties:
               allowedNetworks:
-                description: AllowedNetworks defines which networks are allowed to
-                  connect to, and allow incoming traffic from. The node network is
-                  always allowed.
+                description: AllowedNetworks defines dedicated networks for which
+                  the firewall allows in- and outgoing traffic. The firewall-controller
+                  only enforces this setting in combination with NetworkAccessType
+                  set to forbidden. The node network is always allowed.
                 properties:
                   egress:
-                    description: Egress defines a list of networks which are allowed
-                      for outgoing traffic
+                    description: Egress defines a list of cidrs which are allowed
+                      for outgoing traffic.
                     items:
                       type: string
                     type: array
                   ingress:
-                    description: Ingress defines a list of networks which are allowed
-                      for incoming traffic like service type loadbalancer
+                    description: Ingress defines a list of cidrs which are allowed
+                      for incoming traffic like service type loadbalancer.
                     items:
                       type: string
                     type: array
@@ -149,10 +150,6 @@ spec:
                 description: LogAcceptedConnections if set to true, also log accepted
                   connections in the droptailer log.
                 type: boolean
-              networkAccessType:
-                description: NetworkAccessType defines how the cluster can reach external
-                  networks.
-                type: string
               networks:
                 description: Networks are the networks to which this firewall is connected.
                   An update on this field requires the recreation of the physical
