+++ pkg/apis/metal/types_cloudprofile.go
@@ -72,21 +72,7 @@ type Partition struct {
 	NetworkIsolation *NetworkIsolation
 }
 
-// +k8s:deepcopy-gen:interfaces=k8s.io/apimachinery/pkg/runtime.Object
-
-// ImageProviderConfig is stored in gardenv1beta1.Worker.Image.ProviderConfig as RawExtension
-// to configure metal specific worker configuration items
-type ImageProviderConfig struct {
-	// required to convert it to/from RawExtension
-	metav1.TypeMeta
-	// NetworkIsolation defines restricted/forbidden networkaccess for worker nodes
-	NetworkIsolation *NetworkIsolation
-}
-
-// TODO: NetworkIsolation has actually two purposes:
-// - this is actually also put into shoot.spec.worker.machine.image.providerconfig, we should introduce a imageProviderConfig which includes the networkIsolation for example
-// - transport the configuration for the gepm to create cwnps for clusters with networkisolation configured
-
+// NetworkIsolation defines configuration for restricted or forbidden clusters.
 type NetworkIsolation struct {
 	// AllowedNetworks is a list of networks which are allowed to connect in restricted or forbidden NetworkIsolated clusters.
 	// if empty all destinations are allowed.
