+++ pkg/apis/metal/types_osc.go
@@ -1,0 +1,12 @@
+package metal
+
+import metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+
+// +k8s:deepcopy-gen:interfaces=k8s.io/apimachinery/pkg/runtime.Object
+// ImageProviderConfig is stored in the OSC's provider config RawExtension
+type ImageProviderConfig struct {
+	// required to convert it to/from RawExtension
+	metav1.TypeMeta
+	// NetworkIsolation defines restricted/forbidden networkaccess for worker nodes
+	NetworkIsolation *NetworkIsolation
+}
